+++ frontend/src/pages/inbounds/form/InboundFormModal.tsx
@@ -24,7 +24,6 @@ import {
   DeleteOutlined,
   MinusOutlined,
   PlusOutlined,
-  ReloadOutlined,
 } from '@ant-design/icons';
 
 import { HttpUtil, NumberFormatter, RandomUtil, SizeFormatter, Wireguard } from '@/utils';
@@ -54,9 +53,6 @@ import {
   Protocols,
   SNIFFING_OPTION,
   TCP_CONGESTION_OPTION,
-  TLS_CIPHER_OPTION,
-  TLS_VERSION_OPTION,
-  USAGE_OPTION,
   UTLS_FINGERPRINT,
 } from '@/schemas/primitives';
 import {
@@ -97,8 +93,8 @@ import {
   WsForm,
   XhttpForm,
 } from './transport';
+import { RealityForm, TlsForm } from './security';
 
-const { TextArea } = Input;
 import { coerceInboundJsonField, type DBInbound } from '@/models/dbinbound';
 import type { NodeRecord } from '@/api/queries/useNodesQuery';
 
@@ -1481,421 +1477,28 @@ export default function InboundFormModal({
         </Form.Item>
       </Form.Item>
 
-      <Form.Item
-        noStyle
-        shouldUpdate={(prev, curr) =>
-          prev.streamSettings?.security !== curr.streamSettings?.security
-        }
-      >
-        {({ getFieldValue }) => {
-          const sec = getFieldValue(['streamSettings', 'security']);
-          if (sec !== 'tls') return null;
-          return (
-            <>
-              <Form.Item name={['streamSettings', 'tlsSettings', 'serverName']} label="SNI">
-                <Input placeholder={t('pages.inbounds.form.serverNameIndication')} />
-              </Form.Item>
-              <Form.Item name={['streamSettings', 'tlsSettings', 'cipherSuites']} label={t('pages.inbounds.form.cipherSuites')}>
-                <Select
-                  options={[
-                    { value: '', label: t('pages.inbounds.form.autoOption') },
-                    ...Object.entries(TLS_CIPHER_OPTION).map(([k, v]) => ({ value: v, label: k })),
-                  ]}
-                />
-              </Form.Item>
-              <Form.Item label={t('pages.inbounds.form.minMaxVersion')}>
-                <Space.Compact block>
-                  <Form.Item name={['streamSettings', 'tlsSettings', 'minVersion']} noStyle>
-                    <Select
-                      style={{ width: '50%' }}
-                      options={Object.values(TLS_VERSION_OPTION).map((v) => ({ value: v, label: v }))}
-                    />
-                  </Form.Item>
-                  <Form.Item name={['streamSettings', 'tlsSettings', 'maxVersion']} noStyle>
-                    <Select
-                      style={{ width: '50%' }}
-                      options={Object.values(TLS_VERSION_OPTION).map((v) => ({ value: v, label: v }))}
-                    />
-                  </Form.Item>
-                </Space.Compact>
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'tlsSettings', 'settings', 'fingerprint']}
-                label="uTLS"
-              >
-                <Select
-                  options={[
-                    { value: '', label: 'None' },
-                    ...Object.values(UTLS_FINGERPRINT).map((fp) => ({ value: fp, label: fp })),
-                  ]}
-                />
-              </Form.Item>
-              <Form.Item name={['streamSettings', 'tlsSettings', 'alpn']} label="ALPN">
-                <Select
-                  mode="multiple"
-                  tokenSeparators={[',']}
-                  style={{ width: '100%' }}
-                  options={Object.values(ALPN_OPTION).map((a) => ({ value: a, label: a }))}
-                />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'tlsSettings', 'rejectUnknownSni']}
-                label={t('pages.inbounds.form.rejectUnknownSni')}
-                valuePropName="checked"
-              >
-                <Switch />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'tlsSettings', 'disableSystemRoot']}
-                label={t('pages.inbounds.form.disableSystemRoot')}
-                valuePropName="checked"
-              >
-                <Switch />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'tlsSettings', 'enableSessionResumption']}
-                label={t('pages.inbounds.form.sessionResumption')}
-                valuePropName="checked"
-              >
-                <Switch />
-              </Form.Item>
+      {security === 'tls' && (
+        <TlsForm
+          saving={saving}
+          setCertFromPanel={setCertFromPanel}
+          clearCertFiles={clearCertFiles}
+          generateRandomPinHash={generateRandomPinHash}
+          getNewEchCert={getNewEchCert}
+          clearEchCert={clearEchCert}
+        />
+      )}
 
-              <Form.List name={['streamSettings', 'tlsSettings', 'certificates']}>
-                {(certFields, { add, remove }) => (
-                  <>
-                    <Form.Item label={t('certificate')}>
-                      <Button
-                        type="primary"
-                        size="small"
-                        onClick={() => add({
-                          useFile: true,
-                          certificateFile: '',
-                          keyFile: '',
-                          certificate: [],
-                          key: [],
-                          oneTimeLoading: false,
-                          usage: 'encipherment',
-                          buildChain: false,
-                        })}
-                      >
-                        <PlusOutlined />
-                      </Button>
-                    </Form.Item>
-                    {certFields.map((certField, idx) => (
-                      <div key={certField.key}>
-                        <Form.Item
-                          name={[certField.name, 'useFile']}
-                          label={`${t('certificate')} ${idx + 1}`}
-                        >
-                          <Radio.Group buttonStyle="solid">
-                            <Radio.Button value={true}>
-                              {t('pages.inbounds.certificatePath')}
-                            </Radio.Button>
-                            <Radio.Button value={false}>
-                              {t('pages.inbounds.certificateContent')}
-                            </Radio.Button>
-                          </Radio.Group>
-                        </Form.Item>
-                        {certFields.length > 1 && (
-                          <Form.Item label=" ">
-                            <Button
-                              size="small"
-                              danger
-                              onClick={() => remove(certField.name)}
-                            >
-                              <MinusOutlined /> {t('remove')}
-                            </Button>
-                          </Form.Item>
-                        )}
-                        <Form.Item
-                          noStyle
-                          shouldUpdate={(prev, curr) =>
-                            prev.streamSettings?.tlsSettings?.certificates?.[certField.name]?.useFile
-                            !== curr.streamSettings?.tlsSettings?.certificates?.[certField.name]?.useFile
-                          }
-                        >
-                          {({ getFieldValue }) => {
-                            const useFile = getFieldValue([
-                              'streamSettings', 'tlsSettings', 'certificates',
-                              certField.name, 'useFile',
-                            ]);
-                            return useFile ? (
-                              <>
-                                <Form.Item
-                                  name={[certField.name, 'certificateFile']}
-                                  label={t('pages.inbounds.publicKey')}
-                                >
-                                  <Input />
-                                </Form.Item>
-                                <Form.Item
-                                  name={[certField.name, 'keyFile']}
-                                  label={t('pages.inbounds.privatekey')}
-                                >
-                                  <Input />
-                                </Form.Item>
-                                <Form.Item label=" ">
-                                  <Space>
-                                    <Button
-                                      type="primary"
-                                      loading={saving}
-                                      onClick={() => setCertFromPanel(certField.name)}
-                                    >
-                                      {t('pages.inbounds.setDefaultCert')}
-                                    </Button>
-                                    <Button danger onClick={() => clearCertFiles(certField.name)}>
-                                      {t('clear')}
-                                    </Button>
-                                  </Space>
-                                </Form.Item>
-                              </>
-                            ) : (
-                              <>
-                                <Form.Item
-                                  name={[certField.name, 'certificate']}
-                                  label={t('pages.inbounds.publicKey')}
-                                  normalize={(v) => typeof v === 'string'
-                                    ? v.split('\n')
-                                    : v}
-                                  getValueProps={(v) => ({
-                                    value: Array.isArray(v) ? v.join('\n') : v,
-                                  })}
-                                >
-                                  <TextArea autoSize={{ minRows: 3, maxRows: 8 }} />
-                                </Form.Item>
-                                <Form.Item
-                                  name={[certField.name, 'key']}
-                                  label={t('pages.inbounds.privatekey')}
-                                  normalize={(v) => typeof v === 'string'
-                                    ? v.split('\n')
-                                    : v}
-                                  getValueProps={(v) => ({
-                                    value: Array.isArray(v) ? v.join('\n') : v,
-                                  })}
-                                >
-                                  <TextArea autoSize={{ minRows: 3, maxRows: 8 }} />
-                                </Form.Item>
-                              </>
-                            );
-                          }}
-                        </Form.Item>
-                        <Form.Item
-                          name={[certField.name, 'oneTimeLoading']}
-                          label={t('pages.inbounds.form.oneTimeLoading')}
-                          valuePropName="checked"
-                        >
-                          <Switch />
-                        </Form.Item>
-                        <Form.Item
-                          name={[certField.name, 'usage']}
-                          label={t('pages.inbounds.form.usageOption')}
-                        >
-                          <Select
-                            style={{ width: '50%' }}
-                            options={Object.values(USAGE_OPTION).map((u) => ({ value: u, label: u }))}
-                          />
-                        </Form.Item>
-                        <Form.Item
-                          noStyle
-                          shouldUpdate={(prev, curr) =>
-                            prev.streamSettings?.tlsSettings?.certificates?.[certField.name]?.usage
-                            !== curr.streamSettings?.tlsSettings?.certificates?.[certField.name]?.usage
-                          }
-                        >
-                          {({ getFieldValue }) => {
-                            const usage = getFieldValue([
-                              'streamSettings', 'tlsSettings', 'certificates',
-                              certField.name, 'usage',
-                            ]);
-                            if (usage !== 'issue') return null;
-                            return (
-                              <Form.Item
-                                name={[certField.name, 'buildChain']}
-                                label={t('pages.inbounds.form.buildChain')}
-                                valuePropName="checked"
-                              >
-                                <Switch />
-                              </Form.Item>
-                            );
-                          }}
-                        </Form.Item>
-                      </div>
-                    ))}
-                  </>
-                )}
-              </Form.List>
-
-              <Form.Item name={['streamSettings', 'tlsSettings', 'echServerKeys']} label={t('pages.inbounds.form.echKey')}>
-                <Input />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'tlsSettings', 'settings', 'echConfigList']}
-                label={t('pages.inbounds.form.echConfig')}
-              >
-                <Input />
-              </Form.Item>
-              <Form.Item
-                label={t('pages.inbounds.form.pinnedPeerCertSha256')}
-                tooltip={t('pages.inbounds.form.pinnedPeerCertSha256Tip')}
-              >
-                <Space.Compact block>
-                  <Form.Item
-                    name={['streamSettings', 'tlsSettings', 'settings', 'pinnedPeerCertSha256']}
-                    noStyle
-                  >
-                    <Select
-                      mode="tags"
-                      tokenSeparators={[',', ' ']}
-                      placeholder={t('pages.inbounds.form.pinnedPeerCertSha256Placeholder')}
-                      style={{ width: 'calc(100% - 32px)' }}
-                    />
-                  </Form.Item>
-                  <Button
-                    icon={<ReloadOutlined />}
-                    onClick={generateRandomPinHash}
-                    title={t('pages.inbounds.form.generateRandomPin')}
-                  />
-                </Space.Compact>
-              </Form.Item>
-              <Form.Item label=" ">
-                <Space>
-                  <Button type="primary" loading={saving} onClick={getNewEchCert}>
-                    {t('pages.inbounds.form.getNewEchCert')}
-                  </Button>
-                  <Button danger onClick={clearEchCert}>{t('clear')}</Button>
-                </Space>
-              </Form.Item>
-            </>
-          );
-        }}
-      </Form.Item>
-
-      <Form.Item
-        noStyle
-        shouldUpdate={(prev, curr) =>
-          prev.streamSettings?.security !== curr.streamSettings?.security
-        }
-      >
-        {({ getFieldValue }) => {
-          const sec = getFieldValue(['streamSettings', 'security']);
-          if (sec !== 'reality') return null;
-          return (
-            <>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'show']}
-                label={t('pages.inbounds.form.show')}
-                valuePropName="checked"
-              >
-                <Switch />
-              </Form.Item>
-              <Form.Item name={['streamSettings', 'realitySettings', 'xver']} label={t('pages.inbounds.form.xver')}>
-                <InputNumber min={0} />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'settings', 'fingerprint']}
-                label="uTLS"
-              >
-                <Select
-                  options={Object.values(UTLS_FINGERPRINT).map((fp) => ({ value: fp, label: fp }))}
-                />
-              </Form.Item>
-              <Form.Item label={t('pages.inbounds.form.target')}>
-                <Space.Compact block>
-                  <Form.Item name={['streamSettings', 'realitySettings', 'target']} noStyle>
-                    <Input style={{ width: 'calc(100% - 32px)' }} />
-                  </Form.Item>
-                  <Button icon={<ReloadOutlined />} onClick={randomizeRealityTarget} />
-                </Space.Compact>
-              </Form.Item>
-              <Form.Item label="SNI">
-                <Space.Compact block style={{ display: 'flex' }}>
-                  <Form.Item
-                    name={['streamSettings', 'realitySettings', 'serverNames']}
-                    noStyle
-                  >
-                    <Select mode="tags" tokenSeparators={[',']} style={{ flex: 1 }} />
-                  </Form.Item>
-                  <Button icon={<ReloadOutlined />} onClick={randomizeRealityTarget} />
-                </Space.Compact>
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'maxTimediff']}
-                label={t('pages.inbounds.form.maxTimeDiff')}
-              >
-                <InputNumber min={0} />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'minClientVer']}
-                label={t('pages.inbounds.form.minClientVer')}
-              >
-                <Input placeholder="25.9.11" />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'maxClientVer']}
-                label={t('pages.inbounds.form.maxClientVer')}
-              >
-                <Input placeholder="25.9.11" />
-              </Form.Item>
-              <Form.Item label={t('pages.inbounds.form.shortIds')}>
-                <Space.Compact block style={{ display: 'flex' }}>
-                  <Form.Item
-                    name={['streamSettings', 'realitySettings', 'shortIds']}
-                    noStyle
-                  >
-                    <Select mode="tags" tokenSeparators={[',']} style={{ flex: 1 }} />
-                  </Form.Item>
-                  <Button icon={<ReloadOutlined />} onClick={randomizeShortIds} />
-                </Space.Compact>
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'settings', 'spiderX']}
-                label={t('pages.inbounds.form.spiderX')}
-              >
-                <Input />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'settings', 'publicKey']}
-                label={t('pages.inbounds.publicKey')}
-              >
-                <Input.TextArea autoSize={{ minRows: 1, maxRows: 4 }} />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'privateKey']}
-                label={t('pages.inbounds.privatekey')}
-              >
-                <Input.TextArea autoSize={{ minRows: 1, maxRows: 4 }} />
-              </Form.Item>
-              <Form.Item label=" ">
-                <Space>
-                  <Button type="primary" loading={saving} onClick={genRealityKeypair}>
-                    {t('pages.inbounds.form.getNewCert')}
-                  </Button>
-                  <Button danger onClick={clearRealityKeypair}>{t('clear')}</Button>
-                </Space>
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'mldsa65Seed']}
-                label={t('pages.inbounds.form.mldsa65Seed')}
-              >
-                <Input.TextArea autoSize={{ minRows: 2, maxRows: 6 }} />
-              </Form.Item>
-              <Form.Item
-                name={['streamSettings', 'realitySettings', 'settings', 'mldsa65Verify']}
-                label={t('pages.inbounds.form.mldsa65Verify')}
-              >
-                <Input.TextArea autoSize={{ minRows: 2, maxRows: 6 }} />
-              </Form.Item>
-              <Form.Item label=" ">
-                <Space>
-                  <Button type="primary" loading={saving} onClick={genMldsa65}>
-                    {t('pages.inbounds.form.getNewSeed')}
-                  </Button>
-                  <Button danger onClick={clearMldsa65}>{t('clear')}</Button>
-                </Space>
-              </Form.Item>
-            </>
-          );
-        }}
-      </Form.Item>
+      {security === 'reality' && (
+        <RealityForm
+          saving={saving}
+          randomizeRealityTarget={randomizeRealityTarget}
+          randomizeShortIds={randomizeShortIds}
+          genRealityKeypair={genRealityKeypair}
+          clearRealityKeypair={clearRealityKeypair}
+          genMldsa65={genMldsa65}
+          clearMldsa65={clearMldsa65}
+        />
+      )}
     </>
   );
