+++ frontend/src/pages/inbounds/form/security/index.ts
@@ -1,0 +1,2 @@
+export { default as TlsForm } from './tls';
+export { default as RealityForm } from './reality';
